backend/pyproject.toml
[project]
name = "mae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-auditlog>=3.3.0",
    "django-cacheops>=7.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-ninja>=1.4.5",
    "django-ninja-extra>=0.30.2",
    "django-ninja-jwt>=5.4.0",
    "django-organizations>=2.5.0",
    "django-safedelete>=1.4.1",
    "psycopg[binary]>=3.2.12",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
]
